Reimplement contextual explanations.
I'm not sure why this is slow and bad, but I can guess! It seems like
MosaicML's weights are not especially good when compared to Meta's LLaMa
weights. Additionally, it seems like there's some sort of contention
issue with generation; eight threads are spawned over four cores, which
might be too many threads.